Stock screener

+++ API/UserCustomIndices/Model/Response/RevenueGrowth.cs
@@ -3,7 +3,7 @@
     public struct RevenueGrowth
     {
         public int TimePeriod;
-        public decimal Upper;
-        public decimal Lower;
+        public double Upper;
+        public double Lower;
     }
 }
